從資料庫抓取網頁設定設定 GB改為小數點 抓取硬碟資料(最後一欄暫定為硬碟名稱)

+++ Models/DiskInfo.cs
@@ -10,8 +10,8 @@ namespace collect_all.Models
         public int Id { get; set; }
         public string DeviceNo { get; set; } = string.Empty;
         public string SlotName { get; set; } = string.Empty;
-        public long TotalCapacityGB { get; set; }
-        public long AvailableCapacityGB { get; set; }
+        public float TotalCapacityGB { get; set; }
+        public float AvailableCapacityGB { get; set; }
         public string DeviceInfoDeviceNo { get; set; } = string.Empty;
     }
 }
