+++ Models/HardwareInfo.cs
@@ -12,8 +12,8 @@ namespace collect_all.Models
         public string DeviceNo { get; set; } = string.Empty;
         public string Processor { get; set; } = string.Empty;
         public string Motherboard { get; set; } = string.Empty;
-        public long MemoryTotalGB { get; set; }
-        public long MemoryAvailableGB { get; set; }
+        public float MemoryTotalGB { get; set; }
+        public float MemoryAvailableGB { get; set; }
         public string IPAddress { get; set; } = string.Empty;
         public DateTime? UpdateDate { get; set; }
         public DateTime? CreateDate { get; set; }
